psxc-unpack v2.12: fixed another long standing bug related to [ and ] in dirnames.

+++ zipscript/include/objects.h
@@ -112,7 +112,7 @@ struct current_user {
 struct current_file {
 	char		name[NAME_MAX];
 	char		unlink[NAME_MAX];
-	unsigned int	speed;
+	double		speed;
 	off_t		size;
 	char		compression_method;
 };
